Update API to provide channel index rather than port index
Summary:
The callbacks for send/receive provide an index. This should be the
channel index. Each signal multiplexes two independent channels (left
and right) so with, for example, 3 ports the channels would be 0/1, 2/3
and 4/5. This seems pretty standard and the API is pretty hard to use
when the callback only gives the port index.

This update changes the master code to provide the channel index and
also adds test to check that the sequence of callbacks within a frame
are maintained (and should be documented).

Reviewers: andrew

Reviewed By: andrew

Differential Revision: http://phabricator.xmos.local:9000/D192