First pass at adding TDM component
Summary:
This commit adds TDM to the library. Note that the API has two versions,
one which is a distributable task that wraps around port outputs and
needs to be called in a timely manner and another which mimics the
callback interface of the I2S components.

Currently the testing is sparse and needs improving.

Reviewers: srinivas, andrew

Reviewed By: andrew

Differential Revision: http://phabricator.xmos.local:9000/D158