Add hidden option to priorize the frame start call back
Summary:
This commit adds a hidden option (enabled via a define in the Makefile)
to prioritize the frame start callback. With this option enabled, the
i2s master component will reduce the number of callbacks is does in the
bitclock output slot that it has to do the frame start callback and
spread the callbacks over the other bitclock output slots.

Reviewers: andrew

Reviewed By: andrew

Differential Revision: http://phabricator.xmos.local:9000/D203