Major code and API rework
This commit changes the API so that:

* The frame_start callback is renamed restart_check and no longer
supplies a timestamp.
* All the tasks use the same interface rather than minor variations on
the same one.

The commit also adds a combined i2s/tdm task and moves all the code into
headers to allow the callbacks to be inlined into the i2s/tdm tasks. The
alternative TDM interface is removed.

The tests have been fixed/extended to cover more cases.

The documentation has also been slightly updated but is note complete.