Backport version 2.4.0 fix 6a953e2 from 11 Jan 2018
'There is an issue with xdoc using wavedrom in that wavedrom can occasionally and randomly create zero length png files.  This then cause the nightly builds to fail and the pdf generation fails.

'The workaround is to manually create the wavedrom png files from the js files and update the doc rst to import the firgure rather than generate the images from wavedrom.'