Update loopback demo, still needs DAC/ADC config
Loopback demo updated to use C API and builds for the XCORE.AI Explorer
board. Does not yet configure the DAC and ADC.

The slave loopback demo has been removed. This looked like more of a
test than a demo, and was overly confusing. Not sure how it even worked,
as the DAC and ADC did not appear to be put into master I2S mode.